+++ SadConsole.Host.SFML/Renderers/ScreenObjectRenderer.cs
@@ -37,6 +37,25 @@ namespace SadConsole.Renderers
         {
             // Draw call for texture
             GameHost.Instance.DrawCalls.Enqueue(new DrawCalls.DrawCallTexture(BackingTexture.Texture, new SFML.System.Vector2i(screen.AbsoluteArea.Position.X, screen.AbsoluteArea.Position.Y)));
+
+            if (screen is IScreenObject screenObject)
+            {
+                foreach (var cursor in screenObject.GetSadComponents<Components.Cursor>())
+                {
+                    if (cursor.IsVisible && screen.Surface.IsValidCell(cursor.Position.X, cursor.Position.Y) && screen.Surface.View.Contains(cursor.Position))
+                    {
+                        var cursorPosition = screen.AbsoluteArea.Position + screen.Font.GetRenderRect(cursor.Position.X - screen.Surface.ViewPosition.X, cursor.Position.Y - screen.Surface.ViewPosition.Y, screen.FontSize).Position;
+
+                        GameHost.Instance.DrawCalls.Enqueue(
+                            new DrawCalls.DrawCallCell(cursor.CursorRenderCell,
+                                                       new SadRogue.Primitives.Rectangle(cursorPosition.X, cursorPosition.Y, screen.FontSize.X, screen.FontSize.Y).ToIntRect(),
+                                                       screen.Font,
+                                                       true
+                                                      )
+                            );
+                    }
+                }
+            }
         }
 
         public virtual void Refresh(ISurfaceRenderData screen, bool force = false)